pyproject.toml
[project]
name = "post-processing-from-scratch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bitsandbytes>=0.49.1",
    "cython-bbox>=0.1.5",
    "einops>=0.8.1",
    "gdown>=5.2.0",
    "lap>=0.5.12",
    "lmdb>=1.7.5",
    "mlflow>=3.8.1",
    "nltk>=3.9.2",
    "pytorch-lightning>=2.6.0",
    "qwen-vl-utils>=0.0.14",
    "supervision>=0.27.0",
    "tensorboard>=2.20.0",
    "timm>=1.0.23",
    "torchreid>=0.2.5",
    "transformers>=4.57.3",
    "yacs>=0.1.8",
]
